+++ service.fussballpopup/addon.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<addon id="service.fussballpopup" name="Fussball Ticker" version="0.3.9" provider-name="L0RE">
+<addon id="service.fussballpopup" name="Fussball Ticker" version="0.3.10" provider-name="L0RE">
     <requires>
         <import addon="xbmc.python" version="2.24.0"/>			       
         <import addon="script.module.feedparser" version="5.1.3"/>
-        <import addon="service.popwindow" version="0.2.6"/>	
+        <import addon="service.popwindow" version="0.2.8"/>	
         <import addon="script.module.dateutil" version="2.5.3"/>  
     </requires>
     <extension point="xbmc.service" library="service.py" start="login" />
@@ -11,14 +11,15 @@
 	    <provides>executable</provides>
     </extension>
     <extension point="xbmc.addon.metadata">
-    <summary lang="en">Shows Football Games Messages Feed in Overlay WIndows</summary>
-    <summary lang="de">Zeigt Fussball SPiele Nachrichten Updates in PupupFensterann</summary>
-    <description lang="en">Shows Fussball Feed in Overlay WIndows</description>
-    <description lang="de">Zeigt Fussball Updates in PupupFensterann</description>
+    <summary lang="en">Shows Football Games Messages Feed in Overlay Windows</summary>
+    <summary lang="de">Zeigt Fussball SPiele Nachrichten Updates in Pupup Fenster</summary>
+    <description lang="en">Shows Fussball Feed in Overlay Windows. Games can be added under Programm Addon</description>
+    <description lang="de">Zeigt Fussball Updates in Pupup Fenster. Unter Programm Addons kann man Spiele eintragen</description>
     <website></website>
     <source>https://github.com/kodinerds/repo/tree/master/service.fussballpopup</source>
     <forum>http://www.kodinerds.net/index.php/Thread/52014-Fussball-Meldungen-via-Popup-um-auf-dem-Laufenden-zu-bleiben/</forum>
     <email>andreas at vogler dot name</email>
+    <license>GNU GENERAL PUBLIC LICENSE. Version 2, June 1991</license>
     <platform>all</platform>
   </extension>
 </addon>
