+++ service.fussballpopup/resources/lib/django/__init__.py
