+++ service.fussballpopup/resources/lib/django/shortcuts.py
