+++ service.fussballpopup/resources/lib/django/utils/deprecation.py
