+++ service.fussballpopup/resources/lib/django/utils/encoding.py
