+++ service.fussballpopup/resources/lib/django/utils/functional.py
